streamlit/requirements.txt
joblib
scikit-learn
serial
pyyaml
websockets
streamlit
